Setup polling, can now control synth by sending and pulling data from DB

+++ database/migrations/2019_11_18_193840_create_synthparams_table.php
@@ -16,7 +16,7 @@ class CreateSynthParamsTable extends Migration
         Schema::create('synthparams', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('param_name');
-            $table->unsignedBigInteger('param_val');
+            $table->float('param_val');
             $table->timestamps();
         });
     }
